# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marketplace API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/{id}/ratings:
    get:
      operationId: getProfileRatings
      tags: [ratings]
      summary: Get ratings for a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile ratings summary and list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileRatings"
    post:
      operationId: submitRating
      tags: [ratings]
      summary: Submit a rating for a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRatingBody"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingItem"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RatingItem:
      type: object
      properties:
        id:
          type: string
        raterId:
          type: string
        ratedId:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        raterName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, raterId, ratedId, rating, createdAt]
    ProfileRatings:
      type: object
      properties:
        averageRating:
          type: ["number", "null"]
        totalRatings:
          type: integer
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/RatingItem"
      required: [totalRatings, ratings]
    SubmitRatingBody:
      type: object
      properties:
        raterId:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required: [raterId, rating]
...