# lib/api-zod/src/generated/types/profileRatings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RatingItem } from "./ratingItem";

export interface ProfileRatings {
  /** @nullable */
  averageRating?: number | null;
  totalRatings: number;
  ratings: RatingItem[];
}