# lib/api-zod/src/generated/types/ratingItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RatingItem {
  id: string;
  raterId: string;
  ratedId: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  raterName?: string | null;
  createdAt: string;
}