# lib/api-zod/src/generated/types/submitRatingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketplace API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitRatingBody {
  raterId: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
}